Convert default to named exports
