+++ packages/ra-core/src/export/ExporterContext.ts
@@ -1,10 +1,8 @@
 import { createContext } from 'react';
 
 import { Exporter } from '../types';
-import defaultExporter from './defaultExporter';
+import { defaultExporter } from './defaultExporter';
 
-const ExporterContext = createContext<Exporter | false>(defaultExporter);
+export const ExporterContext = createContext<Exporter | false>(defaultExporter);
 
 ExporterContext.displayName = 'ExporterContext';
-
-export default ExporterContext;
