+++ packages/ra-core/src/export/defaultExporter.ts
@@ -1,9 +1,7 @@
 import jsonExport from 'jsonexport/dist';
 
-import downloadCSV from './downloadCSV';
+import { downloadCSV } from './downloadCSV';
 import { Exporter } from '../types';
 
-const defaultExporter: Exporter = (data, _, __, resource) =>
+export const defaultExporter: Exporter = (data, _, __, resource) =>
     jsonExport(data, (err, csv) => downloadCSV(csv, resource));
-
-export default defaultExporter;
