+++ packages/ra-core/src/export/fetchRelatedRecords.ts
@@ -1,9 +1,5 @@
-import {
-    RaRecord,
-    Identifier,
-    DataProvider,
-    FetchRelatedRecords,
-} from '../types';
+import { DataProvider, FetchRelatedRecords } from '../types';
+import { getRelatedIds } from './getRelatedIds';
 
 /**
  * Helper function for calling the dataProvider.getMany() method,
@@ -17,7 +13,7 @@ import {
  *         }))
  *     );
  */
-const fetchRelatedRecords =
+export const fetchRelatedRecords =
     (dataProvider: DataProvider): FetchRelatedRecords =>
     (data, field, resource) =>
         dataProvider
@@ -28,40 +24,3 @@ const fetchRelatedRecords =
                     return acc;
                 }, {})
             );
-
-/**
- * Extracts, aggregates and deduplicates the ids of related records
- *
- * @example
- *     const books = [
- *         { id: 1, author_id: 123, title: 'Pride and Prejudice' },
- *         { id: 2, author_id: 123, title: 'Sense and Sensibility' },
- *         { id: 3, author_id: 456, title: 'War and Peace' },
- *     ];
- *     getRelatedIds(books, 'author_id'); => [123, 456]
- *
- * @example
- *     const books = [
- *         { id: 1, tag_ids: [1, 2], title: 'Pride and Prejudice' },
- *         { id: 2, tag_ids: [2, 3], title: 'Sense and Sensibility' },
- *         { id: 3, tag_ids: [4], title: 'War and Peace' },
- *     ];
- *     getRelatedIds(records, 'tag_ids'); => [1, 2, 3, 4]
- *
- * @param {Object[]} records An array of records
- * @param {string} field the identifier of the record field to use
- */
-export const getRelatedIds = (
-    records: RaRecord[],
-    field: string
-): Identifier[] =>
-    Array.from(
-        new Set(
-            records
-                .filter(record => record[field] != null)
-                .map(record => record[field])
-                .reduce((ids, value) => ids.concat(value), [])
-        )
-    );
-
-export default fetchRelatedRecords;
