+++ packages/ra-core/src/export/index.ts
@@ -1,6 +1,4 @@
-import defaultExporter from './defaultExporter';
-import downloadCSV from './downloadCSV';
-import ExporterContext from './ExporterContext';
-import fetchRelatedRecords from './fetchRelatedRecords';
-
-export { defaultExporter, downloadCSV, ExporterContext, fetchRelatedRecords };
+export * from './defaultExporter';
+export * from './downloadCSV';
+export * from './ExporterContext';
+export * from './fetchRelatedRecords';
